UNet/requirements.txt
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
tqdm==4.66.5
matplotlib
pillow
numpy
torchinfo==1.8.0